Add a bar between the clip lines
Displays the selected portion of the video, and highlights between the
clip lines. When you drag a clip line, the selected bar resizes
accordingly. When you drag the bar, the clip lines move along with it
to a different part of the video.

NOTE: the dragging responsiveness is a little laggy right now. Kind of
      janky to be honest. Hopefully we can figure out a way to make
      the connection a bit smoother.